requirements.txt
stockfish-3.28.0
Chessnut-0.4.1